Move `add_diag` to `distributions.util` module.
